fix: v2 audit: https://github.com/yAudit/temple-lending-report/issues/20

+++ protocol/contracts/fakes/v2/TempleDebtTokenTestnetAdmin.sol
@@ -28,11 +28,11 @@ contract TempleDebtTokenTestnetAdmin {
         dUSD.burn(from, amount);
     }
 
-    function setBaseInterestRate(uint256 _rate) external {
+    function setBaseInterestRate(uint96 _rate) external {
         dUSD.setBaseInterestRate(_rate);
     }
 
-    function setRiskPremiumInterestRate(address _debtor, uint256 _rate) external {
+    function setRiskPremiumInterestRate(address _debtor, uint64 _rate) external {
         dUSD.setRiskPremiumInterestRate(_debtor, _rate);
     }
