+++ protocol/contracts/interfaces/v2/ITreasuryPriceIndexOracle.sol
@@ -11,11 +11,11 @@ import { ITempleElevatedAccess } from "contracts/interfaces/v2/access/ITempleEle
  * This rate is updated manually with elevated permissions. The new TPI doesn't take effect until after a cooldown.
  */
 interface ITreasuryPriceIndexOracle is ITempleElevatedAccess {
-    event TreasuryPriceIndexSet(uint256 oldTpi, uint256 newTpi);
-    event TpiCooldownSet(uint256 cooldownSecs);
+    event TreasuryPriceIndexSet(uint96 oldTpi, uint96 newTpi);
+    event TpiCooldownSet(uint32 cooldownSecs);
     event MaxTreasuryPriceIndexDeltaSet(uint256 maxDelta);
 
-    error BreachedMaxTpiDelta(uint256 oldTpi, uint256 newTpi, uint256 maxDelta);
+    error BreachedMaxTpiDelta(uint96 oldTpi, uint96 newTpi, uint256 maxDelta);
 
     /**
      * @notice The current Treasury Price Index (TPI) value
@@ -42,12 +42,12 @@ interface ITreasuryPriceIndexOracle is ITempleElevatedAccess {
     /**
      * @notice Set the Treasury Price Index (TPI)
      */
-    function setTreasuryPriceIndex(uint256 value) external;
+    function setTreasuryPriceIndex(uint96 value) external;
 
     /**
      * @notice Set the number of seconds to elapse before a new TPI will take effect.
      */
-    function setTpiCooldown(uint256 cooldownSecs) external;
+    function setTpiCooldown(uint32 cooldownSecs) external;
 
     /**
      * @notice Set the maximum allowed TPI change on any single `setTreasuryPriceIndex()`, in absolute terms.
