+++ protocol/test/forge/v2/templeDebtToken/TempleDebtToken.BaseAndDebtorInterest.t.sol
@@ -6,8 +6,8 @@ import { TempleDebtTokenTestBase } from "./TempleDebtToken.Base.t.sol";
 
 /* solhint-disable func-name-mixedcase, contract-name-camelcase, not-rely-on-time */
 contract TempleDebtTokenTestBaseAndDebtorInterest is TempleDebtTokenTestBase {
-    uint256 public aliceInterestRate = 0.02e18;
-    uint256 public bobInterestRate = 0.05e18;
+    uint64 public aliceInterestRate = 0.02e18;
+    uint64 public bobInterestRate = 0.05e18;
 
     function setUp() public {
         _setUp();
@@ -373,7 +373,7 @@ contract TempleDebtTokenTestBaseAndDebtorInterest is TempleDebtTokenTestBase {
         checkDebtor(bob, bobInterestRate, amount, bobExpected.baseShares, 0, startBlockTs + 1 days, bobExpected.balanceOf-1); // balanceOf rounded down
 
         changePrank(executor);
-        uint256 updatedRate = 0.1e18;
+        uint64 updatedRate = 0.1e18;
         dUSD.setRiskPremiumInterestRate(alice, updatedRate);
 
         // The rate was updated and a checkpoint was made.
@@ -448,7 +448,7 @@ contract TempleDebtTokenTestBaseAndDebtorInterest is TempleDebtTokenTestBase {
         checkDebtor(bob, bobInterestRate, amount, bobExpected.baseShares, 0, startBlockTs + 1 days, bobExpected.balanceOf-1); // balanceOf rounded down
 
         changePrank(executor);
-        uint256 updatedRate = 0;
+        uint64 updatedRate = 0;
         dUSD.setRiskPremiumInterestRate(alice, updatedRate);
 
         // The rate was updated and a checkpoint was made.
